rename load_font.js to fonts.js
